feat(seo,workshops): PR #3 — sitemap Library+OS deep-dives + workshop server/client split + schemas
Closes the SEO surface gap on the highest-revenue product line (workshops)
and surfaces the Library OS + FrankX OS deep-dives to crawlers.

## Workshop server/client split (app/workshops/[slug]/)

Was: page.tsx with 'use client' from line 1. Next.js silently drops metadata
exports from client components, so every paid 8-week workshop served the
parent layout's title/description on social previews + SERP. The Course
schema rendered in JSX was effectively invisible to crawlers because of
hydration timing.

Now:
- page.tsx (server component) — generateStaticParams pre-renders all 8 slugs,
  generateMetadata returns proper title/description/OG with dynamic /api/og
  image card per workshop, three JSON-LD graphs (Course + EducationEvent +
  BreadcrumbList) emit at SSR.
- WorkshopClient.tsx (NEW) — all interactive UI (motion accordions, share
  copy, instructor notes) moved here, accepts `workshop: Workshop` prop.

Schema improvements:
- Course schema: instructor sameAs (LinkedIn + GitHub from social-links SSOT),
  jobTitle 'AI Architect', teaches[] from objectives, coursePrerequisites,
  audience, CourseInstance with mixed mode + workload.
- EducationEvent schema (NEW): VirtualLocation, organizer, performer,
  EducationalAudience, eventAttendanceMode = MixedEventAttendanceMode.
- BreadcrumbList schema (NEW): Home → Workshops → {Workshop title}.

## Sitemap completeness (app/sitemap.ts)

Added 4 fixed pages + 2 dynamic registries:
- /library, /library/approach, /library/build, /library/quotes (fixed)
- /library/{slug} (dynamic, from data/book-reviews.ts — currently ~10 books)
- /os (fixed)
- /os/{slug} (dynamic, from data/os-modules.ts — Watch, Workshop, ACO, ACOS,
  CoE Hub, Library OS, Build First Funnel)

## social-links.ts socialLinks accessor (lib/social-links.ts)

Added flat-object `socialLinks` export deriving from existing SOCIAL_PROFILES.
Ergonomic shape for url-only consumers (schema sameAs, JSX href, email
template interpolation). Mirrors twitter/x/linkedin/github/youtube/instagram/suno.
Doesn't break existing PRIMARY_SOCIAL_LINKS / SCHEMA_SAME_AS / getSocialLink
consumers.

## Net SEO impact

8 paid workshops × 3 JSON-LD graphs each = 24 new structured-data nodes
indexable by Google + AI engines. Plus library + os deep-dives become
discoverable via sitemap (10 books + 7 OS modules = 17 new URLs).

Source: frankxai/FrankX@7c49e878

Refs: docs/ops/2026-05-06-SEO-AEO-AUDIT.md (P0 #1 — highest revenue-blocking)
Refs: docs/ops/2026-05-06-CONVERSION-AUDIT.md (P0 #1)
